embedding_pipeline/requirements.txt
optimum[onnxruntime]>=1.17.0
onnxruntime>=1.17.0
transformers>=4.38.0
qdrant-client>=1.9.0
asyncpg==0.29.0
python-dotenv==1.0.1
tqdm==4.66.4
numpy==1.26.4
